requirements.txt
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-snowflake==3.2.0